Fix running Jest while using ES modules
I had to fix some config values per the Jest documentation.
The recommendation of using ts-jest-resolver was buried in an
issue, but it helped. Finally, I have no idea why I suddenly
needed to install p-try, which is like a sub-sub-sub-dependency
of Jest, but fixed the "module not found" error.